# Stores the registry small strings inline and avoid to use Optional