Two main changes:
* Uses inline strings for most subtags.
* Split the `Record` structs into multiple arrays. Most of the `Optional` fields where empty. It allows us to save some binary size (something like 0.1MB) and improve slightly the speed.

The benchmark binary size without compression decreased from 3.4 Mo to 2.6 Mo and it significantly improved benchmarks:

Before:
```
test canonicalize_tags ... bench: 33,444 ns/iter (+/- 3,792) = 18 MB/s
test parse_well_formed_tags ... bench: 8,566 ns/iter (+/- 367) = 68 MB/s
test validate_tags ... bench: 20,215 ns/iter (+/- 3,330) = 30 MB/s
```

After inline strings but before `Record` split:
```
test canonicalize_tags ... bench: 22,391 ns/iter (+/- 923) = 27 MB/s
test parse_well_formed_tags ... bench: 8,595 ns/iter (+/- 779) = 68 MB/s
test validate_tags ... bench: 11,121 ns/iter (+/- 779) = 56 MB/s
```

Final:
```
test canonicalize_tags      ... bench:      17,915 ns/iter (+/- 1,065) = 34 MB/s
test parse_well_formed_tags ... bench:       8,577 ns/iter (+/- 340) = 68 MB/s
test validate_tags          ... bench:      10,819 ns/iter (+/- 370) = 57 MB/s
```


Better const expressions and const generics will hopefully allow us to merge together the *Subtag structs in the future in a nice way.